requirements.txt
Flask==2.2.2
Pillow==9.2.0
scipy==1.8.0
torch==1.12.1
torchvision==0.13.1
transformers==4.21.3
